tidy up code in subsetting classes #98
removes some debugging code in the cmap replacer and moves the main glyph parsing logic in the glyph table subsetter to a method.

in the next commit we will delete the cmap replacer since it doesn't work and is no longer needed but we want a clean version of it in the commit history for reference.