Don't try to deduplicate a token
Remove the ability to check weather a token that we are writing is duplicated

I decided against checking the token that we are writing with the one that we already have written because for large document it consume too much time.

Make the PdfStreamWriter public so external tool can  use it.