Fix related to page sizes / rotation / coordinate transformations (issue 560):
The initial transformation matrix was incorrect, as it translated by the cropbox width/height
instead of by the cropbox left/bottom offsets. Also, it did not translate the results back into
the 1st quadrant so that (0,0) would (again) be the lower left corner origin for the cropped area.
Added unit tests in new file ContentStreamProcessorTests.

EFFECTIVE CHANGES:

- The coordinates used for letters etc. are different now for rotated and/or cropped pages,
  but as those were not very consistent anyway this is probably OK.

- The Page Size (A4, A3, Custom, etc.), Width and Height are now determined by the CropBox,
  not by the MediaBox; the CropBox ultimately determines what you see on screen and is printable.
  If no cropbox is defined in the PDF, it is set to the MediaBox; so in that case it is
  backwards compatible with the old code.

- The Page MediaBox and CropBox properties are no longer rotated according to Page.Rotation.
  The Page Width and Height do take rotation into account (kept it backward compatible).